requirements.txt
# Document Processing (kept from original)
reportlab
fuzzywuzzy
pytesseract
pdf2image
Pillow
python-Levenshtein>=0.21.0
python-magic>=0.4.27
python-docx>=1.0.0

# LLM (only for reranking, NOT embeddings)
openai>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langsmith>=0.1.0

# Embeddings (HuggingFace — free, biomedical)
sentence-transformers>=2.2.0
chromadb>=0.4.22

# Agent Framework
langgraph>=0.1.0

# Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# PDF Generation
pymupdf>=1.23.0